pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logistics-data-platform"
version = "1.0.0"
description = "Enterprise-grade, event-driven logistics data platform on GCP"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Data Engineering Team", email = "data-engineering@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Apache Beam and Dataflow
    "apache-beam[gcp]>=2.50.0",

    # Google Cloud
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-pubsub>=2.0.0",
    "google-cloud-secret-manager>=2.0.0",
    "google-cloud-monitoring>=2.0.0",

    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jsonschema>=4.0.0",

    # Configuration
    "pyyaml>=6.0.0",

    # Web framework (API)
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",

    # Utilities
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",  # For FastAPI testing

    # Linting and formatting
    "ruff>=0.1.0",
    "mypy>=1.5.0",

    # Type stubs
    "types-PyYAML>=6.0.0",
    "types-jsonschema>=4.0.0",
]

airflow = [
    # Airflow dependencies (for local development)
    "apache-airflow>=2.6.0",
    "apache-airflow-providers-google>=10.0.0",
]

all = [
    "logistics-data-platform[dev,airflow]",
]

[project.scripts]
run-pipeline = "src.streaming.pipeline:run"
run-api = "src.api.main:main"

[project.urls]
Homepage = "https://github.com/example/logistics-data-platform"
Documentation = "https://github.com/example/logistics-data-platform#readme"
Repository = "https://github.com/example/logistics-data-platform"
Issues = "https://github.com/example/logistics-data-platform/issues"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "smoke: Smoke tests for quick validation",
    "slow: Tests that take a long time to run",
    "integration: Integration tests requiring GCP",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "build/",
]
